Изменения в именах файлах и их расположении в проекте Убраны лишние пробелы, импорты ненужных библиотек Изменен состав контроллеров главного экрана
