+++ MyCar/App/Entities/AutoModel.swift
@@ -6,8 +6,8 @@
 //
 
 import UIKit
+
 enum AutoModel: Int, CustomStringConvertible, CaseIterable {
-    
     case item
     case model
     case number
@@ -38,21 +38,5 @@ enum AutoModel: Int, CustomStringConvertible, CaseIterable {
         case .fuelType: return UIImage(named: "penwriting") ?? UIImage()
         case .vin: return UIImage(named: "penwriting") ?? UIImage()
         }
-    }
-}
-struct CarModel {
-    var item: String
-    var model: String
-    var number: String
-    var year: String
-    var distance: Float
-    var vin: String
-    var activFlag: Bool
-    var fuelType: String
-    
-    static func fetchPass() -> [CarModel] {
-        
-        let firstItem = CarModel(item: "", model: "", number: "", year: "", distance: 0, vin: "", activFlag: false, fuelType: "")
-        return [firstItem]
     }
 }
