+++ MyCar/App/Flows/Scenes/AddAuto/AutoCell.swift
@@ -39,7 +39,6 @@ class AutoCell: UITableViewCell {
         return textField
     }()
     
-    
     override init(style: UITableViewCell.CellStyle, reuseIdentifier: String?) {
         super.init(style: style, reuseIdentifier: reuseIdentifier)
         setupElements()
