+++ MyCar/App/Flows/Scenes/FuelPopOver/PopoverCell.swift
@@ -8,10 +8,7 @@
 import UIKit
 
 class PopoverCell: UITableViewCell {
-  
     override func setSelected(_ selected: Bool, animated: Bool) {
         super.setSelected(selected, animated: animated)
-
     }
-
 }
