+++ MyCar/App/Flows/Scenes/Home/HomeViewController.swift
@@ -5,7 +5,6 @@
 //  Created by Aksilont on 06.05.2022.
 //
 
-import UIKit
 import SwiftUI
 import Combine
