+++ MyCar/App/Flows/Scenes/Start/StartViewController.swift
@@ -16,6 +16,4 @@ class StartViewController: UIViewController {
         navigationController?.pushViewController(loginViewController, animated: false)
     }
 
-
 }
-
