+++ MyCar/Core/Extensions/UIApplication+endEditing.swift
@@ -1,5 +1,5 @@
 //
-//  UIApplication+Ext.swift
+//  UIApplication+endEditing.swift
 //  MyCar
 //
 //  Created by Aksilont on 12.05.2022.
