+++ MyCar/Core/Extensions/UIColor+Additions.swift
@@ -1,16 +1,15 @@
 //
-//  UIColor + Extension.swift
+//  UIColor+Additions.swift
 //  MyCar
 //
 //  Created by Alex Larin on 21.05.2022.
 //
 
-import Foundation
 import UIKit
+
 extension UIColor {
     static let checkMarkGray = UIColor(red: 170/255, green: 170/255, blue: 170/255, alpha: 1)
     static let buttonColor = UIColor(red: 54/255, green: 155/255, blue: 161/255, alpha: 1)
     static let checkMarkBlue = UIColor(red: 0.078, green: 0.435, blue: 0.875, alpha: 1)
     static let checkMarkClear = UIColor(red: 1, green: 1, blue: 1, alpha: 0.6)
-    
 }
