+++ MyCar/Core/Extensions/UIHostingControllerCustom.swift
@@ -5,7 +5,6 @@
 //  Created by Aksilont on 12.05.2022.
 //
 
-import UIKit
 import SwiftUI
 
 class UIHostingViewControllerCustom<Content>: UIHostingController<Content> where Content: View {
